Conflicts:
	doc/devel/mainpage.dox
	src/bin/dhcp4/dhcp4.dox
	src/bin/dhcp4/dhcp4_srv.cc
	src/bin/dhcp4/tests/dhcp4_srv_unittest.cc
	src/bin/dhcp4/tests/dhcp4_test_utils.h
	src/lib/dhcp_ddns/tests/ncr_unittests.cc
	src/lib/dhcpsrv/lease_mgr.cc
	src/lib/dhcpsrv/lease_mgr.h
	src/lib/dhcpsrv/tests/lease_mgr_unittest.cc

+++ src/lib/asiolink/tests/tcp_endpoint_unittest.cc
@@ -37,7 +37,7 @@ TEST(TCPEndpointTest, v4Address) {
 
     EXPECT_TRUE(address == endpoint.getAddress());
     EXPECT_EQ(test_port, endpoint.getPort());
-    EXPECT_EQ(IPPROTO_TCP, endpoint.getProtocol());
+    EXPECT_EQ(static_cast<short>(IPPROTO_TCP), endpoint.getProtocol());
     EXPECT_EQ(AF_INET, endpoint.getFamily());
 }
 
@@ -50,6 +50,6 @@ TEST(TCPEndpointTest, v6Address) {
 
     EXPECT_TRUE(address == endpoint.getAddress());
     EXPECT_EQ(test_port, endpoint.getPort());
-    EXPECT_EQ(IPPROTO_TCP, endpoint.getProtocol());
+    EXPECT_EQ(static_cast<short>(IPPROTO_TCP), endpoint.getProtocol());
     EXPECT_EQ(AF_INET6, endpoint.getFamily());
 }
